Applying Multilayer Perceptron on MNIST dataset
Digit recognition is one of the earliest researched areas in machine learning. The handwritten digit recognition is the ability of computers to recognize human handwritten digits. It is a hard task for the machine because handwritten digits are not perfect and can be made with many different flavors. The handwritten digit recognition is the solution to this problem which uses the image of a digit and recognizes the digit present in the image. 
MNIST dataset is one of the most widely used dataset in machine learning digit recognition tasks. It is a collection of 60000 28x28 tiny black and white handwritten images of numeric values from 0 to 9.
This dataset is used in this study to train a multilayer perceptron and build a model that recognizes images with highest accuracy